+++ examples/sveltekit-magic-link/src/routes/dashboard/+page.ts
@@ -1,0 +1,15 @@
+import type { PageLoad } from './$types';
+import { getSupabase } from '@supabase/auth-helpers-sveltekit';
+import { redirect } from '@sveltejs/kit';
+
+export const load: PageLoad = async (event) => {
+	const { session, supabaseClient } = await getSupabase(event);
+	if (!session) {
+		throw redirect(303, '/');
+	}
+	const { data: testTable } = await supabaseClient.from('test').select('*');
+	return {
+		testTable,
+		user: session.user
+	};
+};