+++ packages/svelte/src/index.ts
@@ -1,4 +1,0 @@
-export { type Session, checkSession } from './helpers';
-export { isLoading, error } from './store';
-
-export { default as SupaAuthHelper } from './SupaAuthHelper.svelte';