+++ packages/sveltekit/src/app.d.ts
@@ -1,9 +1,16 @@
-/// <reference types="@sveltejs/kit" />
-
-// See https://kit.svelte.dev/docs/typescript
+// See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
+// and what to do when importing types
 declare namespace App {
-	type AuthSession = import('./types').Locals
-	interface Locals extends AuthSession {
-	}
+  interface Supabase {
+    Database: any;
+    SchemaName: 'public';
+  }
+
+  // interface Locals {}
+  interface PageData {
+    session: import('@supabase/supabase-js').Session | null;
+  }
+  // interface Error {}
+  // interface Platform {}
 }