+++ packages/sveltekit/src/index.ts
@@ -1,18 +1,4 @@
-// Types
-export type { User, SupabaseClient } from '@supabase/supabase-js';
-
-// Methods
-export {
-  handleSession,
-  handleCallback,
-  handleUser,
-  handleLogout,
-  handleAuth
-} from './handlers/index';
-export { default as getUserAndSaveTokens, getUser } from './utils/getUser';
-export { getProviderToken } from './utils/getProviderToken';
-export { default as withApiAuth } from './utils/withApiAuth';
-export { default as withPageAuth } from './utils/withPageAuth';
-export { default as supabaseServerClient } from './utils/supabaseServerClient';
-export { skHelper, createSupabaseClient } from './instance';
-export { default as logger } from './utils/log';
+export type { ExtendedEvent, Config } from './types';
+export { getSupabase } from './utils/getSupabase';
+export { getServerSession } from './utils/getServerSession';
+export { createClient } from './createClient';