+++ packages/sveltekit/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @supabase/auth-helpers-sveltekit
 
+## 0.8.2
+
+### Patch Changes
+
+- 8f6a474: Fix type issue in the getServerSession function
+
+## 0.8.1
+
+### Patch Changes
+
+- 7a6377a: Add fix for esm package build
+- 7a6377a: Remove old files from 0.7.x
+
+## 0.8.0
+
+### Minor Changes
+
+- 58e6287: Update to supabase-js v2
+
+## 0.7.1
+
+### Patch Changes
+
+- 3a09401: Add type intellisense for @supabase/auth-helpers-sveltekit/server
+  Add .js to fix ERR_MODULE_NOT_FOUND (can´t use the script, it throws an error @sveltejs/kit/hooks not found)
+  Pass page store and invalidation to startSupabaseSessionSync as \$app module is not available in npm packages
+
+## 0.7.0
+
+### Minor Changes
+
+- cd9150c: [breaking] Update to work with latest SvelteKit RC
+
+## 0.6.11
+
+### Patch Changes
+
+- Updated dependencies [63b1da0]
+  - @supabase/auth-helpers-shared@0.1.4
+
 ## 0.6.10
 
 ### Patch Changes